third_party/WebKit/LayoutTests/accessibility/inert-node-is-hidden-expected.txt

This tests that inert elements are not in the AT tree.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS accessibilityController.accessibleElementById('input') is non-null.
PASS treeContainsText(webArea, 'This content should not be exposed to AT when a modal dialog is open') is true
PASS accessibilityController.accessibleElementById('input') is null
PASS treeContainsText(webArea, 'This content should not be exposed to AT when a modal dialog is open') is false
PASS accessibilityController.accessibleElementById('input') is non-null.
PASS treeContainsText(webArea, 'This content should not be exposed to AT when a modal dialog is open') is true
PASS successfullyParsed is true

TEST COMPLETE
This content should not be exposed to AT when a modal dialog is open.
